# docs/swagger.yaml
basePath: /api/
definitions:
...
  10_0_0_50_tuan_quang_tran_vms-v2_internal_models.CdnFile:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      file_id:
        type: string
      index:
        type: integer
      offset:
        type: integer
      size:
        type: integer
      type:
        type: string
    type: object
...
  10_0_0_50_tuan_quang_tran_vms-v2_internal_models.WaterMarkFile:
    properties:
      file:
        $ref: '#/definitions/10_0_0_50_tuan_quang_tran_vms-v2_internal_models.CdnFile'
      file_id:
        type: string
      water_mark_id:
        type: string
    type: object
  10_0_0_50_tuan_quang_tran_vms-v2_internal_models.Watermark:
    properties:
      created_at:
        type: string
      height:
        type: integer
      id:
        type: string
      media_files:
        $ref: '#/definitions/10_0_0_50_tuan_quang_tran_vms-v2_internal_models.WaterMarkFile'
      updated_at:
        type: string
      user_id:
        type: string
      watermark_name:
        type: string
      width:
        type: integer
    type: object
...
  CreateWatermarkData:
    properties:
      created_at:
        type: string
      watermark_id:
        type: string
    type: object
  CreateWatermarkResponse:
    properties:
      data:
        $ref: '#/definitions/CreateWatermarkData'
      status:
        type: string
    type: object
...
  GetAllWatermarkData:
    properties:
      total:
        type: integer
      watermarks:
        items:
          $ref: '#/definitions/10_0_0_50_tuan_quang_tran_vms-v2_internal_models.Watermark'
        type: array
    type: object
  GetAllWatermarkResponse:
    properties:
      data:
        $ref: '#/definitions/GetAllWatermarkData'
      status:
        type: string
    type: object
...
  GetWebhooksListData:
    properties:
      total:
        type: integer
      webhooks:
        items:
          $ref: '#/definitions/Webhook'
        type: array
    type: object
  GetWebhooksListResponse:
    properties:
      data:
        $ref: '#/definitions/GetWebhooksListData'
      status:
        type: string
    type: object
...
  ResponseError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  Webhook:
    properties:
      created_at:
        type: string
      encoding_failed:
        type: boolean
      encoding_finished:
        type: boolean
      encoding_started:
        type: boolean
      file_received:
        type: boolean
      id:
        type: string
      last_triggered_at:
        type: string
      name:
        type: string
      partial_finished:
        type: boolean
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aioz Stream Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aioz Stream API
  version: "1.0"
paths:
...
  /watermarks:
    get:
      consumes:
        - application/json
      description: List all watermarks associated with your workspace.
      operationId: LIST-watermarks
      parameters:
        - default: created_at
          description: sort by
          enum:
            - created_at
            - name
          in: query
          name: sort_by
          type: string
        - default: asc
          description: 'allowed: asc, desc. Default: asc'
          enum:
            - asc
            - desc
          in: query
          name: order_by
          type: string
        - default: 0
          description: offset, allowed values greater than or equal to 0. Default(0)
          in: query
          minimum: 0
          name: offset
          type: integer
        - default: 25
          description: results per page. Allowed values 1-100, default is 25
          in: query
          maximum: 100
          minimum: 1
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/GetAllWatermarkResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: List all watermarks
      tags:
        - watermark
      x-client-action: list
      x-client-paginated: true
      x-group-parameters: true
    post:
      consumes:
        - multipart/form-data
      description: Create a new watermark by uploading a JPG or a PNG image.
      operationId: POST_watermark
      parameters:
        - description: Watermark image file
          format: jpg, jpeg, png
          in: formData
          name: file
          required: true
          type: file
      produces:
        - application/json
      responses:
        "201":
          description: Created
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/CreateWatermarkResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Create a new watermark
      tags:
        - watermark
      x-client-action: upload
  /watermarks/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a watermark.
      operationId: DELETE_watermark
      parameters:
        - description: Watermark ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Get list webhooks
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/GetWebhooksListResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Delete a watermark by ID
      tags:
        - watermark
      x-client-action: delete
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"